SOLR-434 - change interfaces should support >2B docs
git-svn-id: https://svn.apache.org/repos/asf/lucene/solr/trunk@610414 13f79535-47bb-0310-9956-ffa450edef68

+++ client/java/solrj/src/org/apache/solr/client/solrj/impl/XMLResponseParser.java
@@ -313,10 +313,10 @@ public class XMLResponseParser implements ResponseParser
       String n = parser.getAttributeLocalName( i );
       String v = parser.getAttributeValue( i );
       if( "numFound".equals( n ) ) {
-        docs.setNumFound( Integer.parseInt( v ) );
+        docs.setNumFound( Long.parseLong( v ) );
       }
       else if( "start".equals( n ) ) {
-        docs.setStart( Integer.parseInt( v ) );
+        docs.setStart( Long.parseLong( v ) );
       }
       else if( "maxScore".equals( n ) ) {
         docs.setMaxScore( Float.parseFloat( v ) );
